CPU and memory leak improvements
The big fix is to remove the use of Docker::Messages for parsing the log
stream.  Whilst it means we've got some more code to maintain, it has the
decided benefit of not caching EVERY SINGLE LOG LINE EVER in a giant,
never-cleared array which we'll never actually get to see.  If I sound
bitter, it's probably because I am.

I also took the liberty of not parsing timestamps when I don't have to,
which seemed like a useful improvement.