Stop Excon trying to be helpful by retrying timed-out log retrievals
Excon has this feature which makes it retry requests when they time out (or,
I presume, fail in other ways).  Now, of course, we all know that some HTTP
requests can't be retried safely (*cough*POST*cough*) because they're not
*idempotent*.  But Excon knows this, and only retries requests which it
knows are (*cough*should be*cough*) idempotent.

Technically, you could argue that Excon is doing the right thing here, too,
because by the HTTP definition of idempotency (which is all about state
change *on the server*), the "give me all the logs since date X" requests
are, indeed, idempotent.  However, when it comes to progressive processing
like that which you get from Docker's approach to streaming logs, all the
usual guarantees about whether a request is safe to retry go out the window.

I guess this is a long-winded way of saying that Docker has, once again, by
it's appalling approach to API design, made a mockery of HTTP, and I would
be extremely happy to fire this entire thing into the sun.