Stop re-reading the last log message on every docker timeout
The Docker logs endpoint uses the parameter "since" to accept a Unix
timestamp to specify a time from which logs will be returned.
Unfortunately, Docker was built by people who lack a firm grasp of the
English language (or perhaps floating-point arithmetic, I guess), and so
when you ask for all logs "since" the timestamp of the last log entry you
previously saw, you get the last log entry again.

So now we add one nanosecond to the timestamp which we pass to "since",
because otherwise we don't get log entries *since* that time, but rather we
get log entries since or *at* that time.