Push an immutable tag on release
This commit will allow old mobystash container images to linger on
Docker Hub after a new release.

https://hub.docker.com/r/discourse/mobystash/tags/

In turn, this will allow users to pin to a specific image hash (and be
confident that it will not disappear unexpectedly), and roll back to an
earlier mobystash release in the event of a regression.  The mutable
`:latest` tag behaves as it did before.

This version number is the simplest thing that could possibly work.  It
is syntactically compliant with semver.org, though these version numbers
are not chronologically or semantically comparable.  For now, we only
require a unique ref per release.  If necessary, semantics may be
trivially introduced in the future by bumping the MINOR version.