Update gems and image to Ruby 2.6 + jemalloc
This adds support for jemalloc, in turn it means the alpine image is a no
go so we switched to the strech slim ruby image.

Dockerfile contains horrendous hack (repeat bundle install) for the life
of me I can not figure out why it is required.

Switch travis to test on 2.5 and 2.6

Update all dependencies in lock file